address review feedback: fix pagination focus, move DonationViewModal, remove vestigial selection state
